Fix validation messages and move to separate files
