+++ bluebottle/assignments/validators.py
@@ -1,0 +1,16 @@
+from django.utils.translation import ugettext_lazy as _
+
+from bluebottle.utils.models import Validator
+
+
+class RegistrationDeadlineValidator(Validator):
+    field = 'registration_deadline'
+    code = 'registration_deadline'
+    message = _('The registration deadline must be before the end'),
+
+    def is_valid(self):
+        return (
+            not self.instance.registration_deadline or
+            not self.instance.date or
+            self.instance.registration_deadline < self.instance.date.date()
+        )
